+++ actstream/__init__.py
@@ -1,13 +1,13 @@
 try:
     from actstream.signals import action
-except ImportError:
+except:
     pass
 
 __version_info__ = {
     'major': 0,
-    'minor': 4,
-    'micro': 5,
-    'releaselevel': 'beta',
+    'minor': 5,
+    'micro': 1,
+    'releaselevel': 'final',
     'serial': 1
 }
 
@@ -23,3 +23,5 @@ def get_version(release_level=True):
 
 
 __version__ = get_version()
+__author__ = 'Justin Quick <justquick@gmail.com>'
+default_app_config = 'actstream.apps.ActstreamConfig'