+++ actstream/actions.py
@@ -1,10 +1,13 @@
 import datetime
 
 from django.utils.translation import ugettext_lazy as _
+from django.utils.six import text_type
 from django.contrib.contenttypes.models import ContentType
 
-from actstream.exceptions import check_actionable_model
 from actstream import settings
+from actstream.signals import action
+from actstream.registry import check
+from actstream.compat import get_model
 
 try:
     from django.utils import timezone
@@ -13,7 +16,7 @@ except ImportError:
     now = datetime.datetime.now
 
 
-def follow(user, obj, send_action=True, actor_only=True):
+def follow(user, obj, send_action=True, actor_only=True, **kwargs):
     """
     Creates a relationship allowing the object's activities to appear in the
     user's stream.
@@ -22,6 +25,7 @@ def follow(user, obj, send_action=True, actor_only=True):
 
     If ``send_action`` is ``True`` (the default) then a
     ``<user> started following <object>`` action signal is sent.
+    Extra keyword arguments are passed to the action.send call.
 
     If ``actor_only`` is ``True`` (the default) then only actions where the
     object is the actor will appear in the user's activity stream. Set to
@@ -32,16 +36,14 @@ def follow(user, obj, send_action=True, actor_only=True):
 
         follow(request.user, group, actor_only=False)
     """
-    from actstream.models import Follow, action
-
-    check_actionable_model(obj)
-    follow, created = Follow.objects.get_or_create(user=user,
-        object_id=obj.pk,
+    check(obj)
+    instance, created = get_model('actstream', 'follow').objects.get_or_create(
+        user=user, object_id=obj.pk,
         content_type=ContentType.objects.get_for_model(obj),
         actor_only=actor_only)
     if send_action and created:
-        action.send(user, verb=_('started following'), target=obj)
-    return follow
+        action.send(user, verb=_('started following'), target=obj, **kwargs)
+    return instance
 
 
 def unfollow(user, obj, send_action=False):
@@ -55,11 +57,11 @@ def unfollow(user, obj, send_action=False):
 
         unfollow(request.user, other_user)
     """
-    from actstream.models import Follow, action
-
-    check_actionable_model(obj)
-    Follow.objects.filter(user=user, object_id=obj.pk,
-        content_type=ContentType.objects.get_for_model(obj)).delete()
+    check(obj)
+    get_model('actstream', 'follow').objects.filter(
+        user=user, object_id=obj.pk,
+        content_type=ContentType.objects.get_for_model(obj)
+    ).delete()
     if send_action:
         action.send(user, verb=_('stopped following'), target=obj)
 
@@ -74,32 +76,29 @@ def is_following(user, obj):
 
         is_following(request.user, group)
     """
-    from actstream.models import Follow
-
-    check_actionable_model(obj)
-    return bool(Follow.objects.filter(user=user, object_id=obj.pk,
-        content_type=ContentType.objects.get_for_model(obj)).count())
+    check(obj)
+    return get_model('actstream', 'follow').objects.filter(
+        user=user, object_id=obj.pk,
+        content_type=ContentType.objects.get_for_model(obj)
+    ).exists()
 
 
 def action_handler(verb, **kwargs):
     """
     Handler function to create Action instance upon action signal call.
     """
-    from actstream.models import Action
-
     kwargs.pop('signal', None)
     actor = kwargs.pop('sender')
-    check_actionable_model(actor)
 
     # We must store the unstranslated string
     # If verb is an ugettext_lazyed string, fetch the original string
     if hasattr(verb, '_proxy____args'):
         verb = verb._proxy____args[0]
 
-    newaction = Action(
+    newaction = get_model('actstream', 'action')(
         actor_content_type=ContentType.objects.get_for_model(actor),
         actor_object_id=actor.pk,
-        verb=unicode(verb),
+        verb=text_type(verb),
         public=bool(kwargs.pop('public', True)),
         description=kwargs.pop('description', None),
         timestamp=kwargs.pop('timestamp', now())
@@ -107,11 +106,12 @@ def action_handler(verb, **kwargs):
 
     for opt in ('target', 'action_object'):
         obj = kwargs.pop(opt, None)
-        if not obj is None:
-            check_actionable_model(obj)
+        if obj is not None:
+            check(obj)
             setattr(newaction, '%s_object_id' % opt, obj.pk)
             setattr(newaction, '%s_content_type' % opt,
                     ContentType.objects.get_for_model(obj))
     if settings.USE_JSONFIELD and len(kwargs):
         newaction.data = kwargs
-    newaction.save()
+    newaction.save(force_insert=True)
+    return newaction