+++ actstream/views.py
@@ -16,8 +16,9 @@ def respond(request, code):
     Responds to the request with the given response code.
     If ``next`` is in the form, it will redirect instead.
     """
-    if 'next' in request.REQUEST:
-        return HttpResponseRedirect(request.REQUEST['next'])
+    redirect = request.GET.get('next', request.POST.get('next'))
+    if redirect:
+        return HttpResponseRedirect(redirect)
     return type('Response%d' % code, (HttpResponse, ), {'status_code': code})()
 
 
@@ -29,12 +30,12 @@ def follow_unfollow(request, content_type_id, object_id, do_follow=True, actor_o
     actor defined by ``content_type_id``, ``object_id``.
     """
     ctype = get_object_or_404(ContentType, pk=content_type_id)
-    actor = get_object_or_404(ctype.model_class(), pk=object_id)
+    instance = get_object_or_404(ctype.model_class(), pk=object_id)
 
     if do_follow:
-        actions.follow(request.user, actor, actor_only=actor_only)
+        actions.follow(request.user, instance, actor_only=actor_only)
         return respond(request, 201)   # CREATED
-    actions.unfollow(request.user, actor)
+    actions.unfollow(request.user, instance)
     return respond(request, 204)   # NO CONTENT
 
 
@@ -44,7 +45,7 @@ def stream(request):
     Index page for authenticated user's activity stream. (Eg: Your feed at
     github.com)
     """
-    return render_to_response(('actstream/actor.html', 'activity/actor.html'), {
+    return render_to_response('actstream/actor.html', {
         'ctype': ContentType.objects.get_for_model(User),
         'actor': request.user, 'action_list': models.user_stream(request.user)
     }, context_instance=RequestContext(request))
@@ -56,9 +57,9 @@ def followers(request, content_type_id, object_id):
     ``content_type_id``, ``object_id``.
     """
     ctype = get_object_or_404(ContentType, pk=content_type_id)
-    actor = get_object_or_404(ctype.model_class(), pk=object_id)
-    return render_to_response(('actstream/followers.html', 'activity/followers.html'), {
-        'followers': models.followers(actor), 'actor': actor
+    instance = get_object_or_404(ctype.model_class(), pk=object_id)
+    return render_to_response('actstream/followers.html', {
+        'followers': models.followers(instance), 'actor': instance
     }, context_instance=RequestContext(request))
 
 
@@ -66,9 +67,9 @@ def following(request, user_id):
     """
     Returns a list of actors that the user identified by ``user_id`` is following (eg who im following).
     """
-    user = get_object_or_404(User, pk=user_id)
-    return render_to_response(('actstream/following.html', 'activity/following.html'), {
-        'following': models.following(user), 'user': user
+    instance = get_object_or_404(User, pk=user_id)
+    return render_to_response('actstream/following.html', {
+        'following': models.following(instance), 'user': instance
     }, context_instance=RequestContext(request))
 
 
@@ -76,10 +77,10 @@ def user(request, username):
     """
     ``User`` focused activity stream. (Eg: Profile page twitter.com/justquick)
     """
-    user = get_object_or_404(User, username=username, is_active=True)
-    return render_to_response(('actstream/actor.html', 'activity/actor.html'), {
+    instance = get_object_or_404(User, **{'is_active': True, compat.username_field(): username})
+    return render_to_response('actstream/actor.html', {
         'ctype': ContentType.objects.get_for_model(User),
-        'actor': user, 'action_list': models.user_stream(user)
+        'actor': instance, 'action_list': models.user_stream(instance)
     }, context_instance=RequestContext(request))
 
 
@@ -87,7 +88,7 @@ def detail(request, action_id):
     """
     ``Action`` detail view (pretty boring, mainly used for get_absolute_url)
     """
-    return render_to_response(('actstream/detail.html', 'activity/detail.html'), {
+    return render_to_response('actstream/detail.html', {
         'action': get_object_or_404(models.Action, pk=action_id)
     }, context_instance=RequestContext(request))
 
@@ -98,9 +99,9 @@ def actor(request, content_type_id, object_id):
     ``object_id``.
     """
     ctype = get_object_or_404(ContentType, pk=content_type_id)
-    actor = get_object_or_404(ctype.model_class(), pk=object_id)
-    return render_to_response(('actstream/actor.html', 'activity/actor.html'), {
-        'action_list': models.actor_stream(actor), 'actor': actor,
+    instance = get_object_or_404(ctype.model_class(), pk=object_id)
+    return render_to_response('actstream/actor.html', {
+        'action_list': models.actor_stream(instance), 'actor': instance,
         'ctype': ctype
     }, context_instance=RequestContext(request))
 
@@ -111,8 +112,8 @@ def model(request, content_type_id):
     ``object_id``.
     """
     ctype = get_object_or_404(ContentType, pk=content_type_id)
-    actor = ctype.model_class()
-    return render_to_response(('actstream/actor.html', 'activity/actor.html'), {
-        'action_list': models.model_stream(actor), 'ctype': ctype,
-        'actor': actor
+    model_class = ctype.model_class()
+    return render_to_response('actstream/actor.html', {
+        'action_list': models.model_stream(model_class), 'ctype': ctype,
+        'actor': model_class
     }, context_instance=RequestContext(request))